completing pipeline
